+++ docs/source/conf.py
@@ -61,9 +61,6 @@ intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),
     "rich": ("https://rich.readthedocs.io/en/latest", None),
     "torch": ("https://pytorch.org/docs/stable", None),
-    "pytorch_lightning": ("https://pytorch-lightning.readthedocs.io/en/latest", None),
-    # Just leaving this here for when we do finally add a deepspeed integration:
-    #  "deepspeed": ("https://deepspeed.readthedocs.io/en/latest/", None),
     "flax": ("https://flax.readthedocs.io/en/latest", None),
     "fairscale": ("https://fairscale.readthedocs.io/en/latest/", None),
     "datasets": ("https://huggingface.co/docs/datasets/master/en", None),